requirements.txt
APScheduler==3.6.3
attrs==23.1.0
cachetools==4.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
docopt==0.6.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pipreqs==0.4.13
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.4
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.1.0
websockets==12.0
wsproto==1.2.0
yarg==0.1.9
